requirements.txt
dataclasses==0.6
urllib3==1.25.3
requests==2.22.0
typing==3.7.4
dataclasses_json==0.2.14
selenium_wire==1.0.8
beautifulsoup4==4.8.0
python-slugify==5.0.2
PyYAML==5.3.1
setuptools~=45.2.0
Pillow~=8.2.0
termcolor~=1.1.0
